add commit

+++ libs/VS2022/LearnOGL/LearnOGLCustomFBO.cpp
@@ -50,7 +50,7 @@ namespace OGL
 		for (uint32_t i = 0; i < mColorAttchCount; i++)
 		{
 			glBindTexture(GL_TEXTURE_2D, mColorTexs[i]);
-			glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16, width, height, 0, GL_RGBA, GL_FLOAT, NULL);
+			glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, width, height, 0, GL_RGBA, GL_FLOAT, NULL);
 			glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 			glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 			glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
